+++ src/main/java/com/duzy/common/task/ProjectTask.java
@@ -46,14 +46,14 @@ public class ProjectTask {
 
 //    @Scheduled(fixedRate = 6000)
     public void redisRemoveTask() {
-        List<Integer> ids = new ArrayList<>();
+        List<Long> ids = new ArrayList<>();
         Iterable<SysUserModel> all = userRepository.findAll();
         all.forEach(user -> {
             ids.add(user.getId());
             log.info("{}", user);
         });
         ids.stream().limit(ids.size() - 1).forEach(id -> {
-            userRepository.deleteById(id);
+            userRepository.deleteById(id.intValue());
         });
 
     }
